Update Gemfile.lock to for mimemagic and Ruby 3
The gem mimemagic was yanked so the currently listed version will no longer install on new machines. This PR updates to a recent version (MIT license) and updates the gems to be able to work with Ruby 3.

In addition I'm bumping the minimum Ruby version to 2.5. Ruby 2.3 is quite old and has been EOL for some time. The oldest Ruby version on Heroku-18 is currently 2.4 but new apps shouldn't be using it.