Faster nokogiri installs (#103)
With https://nokogiri.org/tutorials/installing_nokogiri.html#how-can-i-avoid-using-a-precompiled-native-gem nokogiri now ships pre-compiled. By locking to the x86 platform we are telling Bundler it's okay to install a platform specific gem. This is used to download and install nokogiri without having to wait for it to compile. This will save on build time.