Schneems/update readme dec 5 2024 (#168)
* Update `heroku local` port

The default is not 5006

* Title-ize the README title

* Consistent paragraph spacing

Single space after the period.

* Link to Ruby on Rails main page

* Update installation language

* Use HTTPs URL

* Remove migration

The migration step won't work as we haven't allocated a database yet. If they need that they should go through the tutorial.

* Update links at the bottom

* Update deploy instructions

* Update code fence language

Both `sh` and `console` change the color of `local` and `exec` which isn't what we want. Use `text` to avoid that.

Use `console` for individual console sessions:

https://github.com/orgs/community/discussions/35615